fix post_process.py

+++ scripts/ASBcalling/post_process.py
@@ -39,11 +39,11 @@ else:
     p_val_ref = []
     p_val_alt = []
 table = table.drop(['fdrp_ref', 'fdrp_alt', 'fdrp_ref_fisher', 'fdrp_alt_fisher'], axis=1)
-fdr_by_ref = np.array(['NaN'] * len(table.index))
+fdr_by_ref = np.array(['NaN'] * len(table.index), dtype=np.float128)
 fdr_by_ref[mc_filter_array] = p_val_ref
 table["fdrp_by_ref"] = fdr_by_ref
 
-fdr_by_alt = np.array(['NaN'] * len(table.index))
+fdr_by_alt = np.array(['NaN'] * len(table.index), dtype=np.float128)
 fdr_by_alt[mc_filter_array] = p_val_alt
 table["fdrp_by_alt"] = fdr_by_alt
